requirements.txt
carla
hydra-core
numpy
opencv-python
pyyaml
tqdm
joblib
